+++ src/Query/QueryManager.php
@@ -7,23 +7,38 @@ namespace Strata\Data\Query;
 use Strata\Data\Collection;
 use Strata\Data\DataProviderInterface;
 use Strata\Data\Exception\QueryManagerException;
+use Strata\Data\Http\Http;
 use Strata\Data\Http\Response\CacheableResponse;
 use Strata\Data\Mapper\MapCollection;
+use Strata\Data\Mapper\MapItem;
 use Strata\Data\Mapper\WildcardMappingStrategy;
+use Strata\Data\Query\BuildQuery\BuildGraphQLQuery;
+use Strata\Data\Query\BuildQuery\BuildQuery;
+use Strata\Data\Query\QueryManager\QueryStack;
+use Strata\Data\Query\QueryManager\StackItem;
+use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
  * Class to help manage running queries against APIs
  */
 class QueryManager
 {
-    /** @var Query[]  */
-    private array $queries = [];
-
     /** @var DataProviderInterface[] */
     private array $dataProviders = [];
 
-    /** @var string[] */
-    private array $dataProvidersForQueries = [];
+    /** @var string */
+    public ?string $lastDataProviderName = null;
+
+    /** @var QueryStack  */
+    private QueryStack $queryStack;
+
+    /**
+     * QueryManager constructor.
+     */
+    public function __construct()
+    {
+        $this->queryStack = new QueryStack();
+    }
 
     /**
      * Add a data provider to use with queries
@@ -33,6 +48,23 @@ class QueryManager
     public function addDataProvider(string $name, DataProviderInterface $dataProvider)
     {
         $this->dataProviders[$name] = $dataProvider;
+        $this->lastDataProviderName = $name;
+    }
+
+    /**
+     * Set HTTP client for all data providers
+     *
+     * Primarily used for mocking the HTTPClient for testing
+     *
+     * @param HttpClientInterface $client
+     */
+    public function setHttpClient(HttpClientInterface $client)
+    {
+        foreach ($this->dataProviders as $dataProvider) {
+            if ($dataProvider instanceof Http) {
+                $dataProvider->setHttpClient($client);
+            }
+        }
     }
 
     /**
@@ -48,67 +80,67 @@ class QueryManager
     /**
      * Return named data provider
      * @param string $name
-     * @return DataProviderInterface|null
+     * @return DataProviderInterface
+     * @throws QueryManagerException
      */
-    public function getDataProvider(string $name): ?DataProviderInterface
+    public function getDataProvider(string $name): DataProviderInterface
     {
-        if ($this->hasDataProvider($name)) {
-            return $this->dataProviders[$name];
+        if (!$this->hasDataProvider($name)) {
+            throw new QueryManagerException(sprintf('Cannot find data provider %s', $name));
         }
-        return null;
+        return $this->dataProviders[$name];
     }
 
     /**
-     * Return data provider to use with a named query
-     * @param string $queryName
-     * @return DataProviderInterface
+     * Get last data provider name
+     * @return string
+     * @throws QueryManagerException
      */
-    public function getDataProviderForQuery(string $queryName): DataProviderInterface
+    public function getLastDataProviderName(): string
     {
-        $dataProvider = null;
-        if (isset($this->dataProvidersForQueries[$queryName])) {
-            $dataProvider = $this->getDataProvider($this->dataProvidersForQueries[$queryName]);
+        if ($this->lastDataProviderName === null) {
+            throw new QueryManagerException('You must set at least one data provider to the query manager');
         }
-
-        if (!($dataProvider instanceof DataProviderInterface)) {
-            throw new QueryManagerException(sprintf('Cannot find data provider for query name %s', $queryName));
-        }
-
-        return $dataProvider;
+        return $this->lastDataProviderName;
     }
 
     /**
      * Add a query (does not run the query, this happens on data access)
      *
-     * @param string $dataProviderName
-     * @param Query $query
-     * @param string|null $queryName
+     * @param Query $query Query
+     * @param string $dataProviderName Data provider to use with query, if not set use last added data provider
      * @throws QueryManagerException
      */
-    public function add(string $dataProviderName, Query $query, ?string $queryName = null)
+    public function add(Query $query, ?string $dataProviderName = null)
     {
-        // Get query name from parameters or query object
-        if ($queryName === null) {
-            $queryName = $query->getName();
+        if (!$query->hasName()) {
+            throw new QueryManagerException('Query must have a name before it is added to the Query Manager');
         }
-        if (array_key_exists($queryName, $this->queries)) {
-            throw new QueryManagerException(sprintf('Cannot add query since query with same name "%s" already exists', $query->getName()));
+        if ($this->queryStack->offsetExists($query->getName())) {
+            throw new QueryManagerException(sprintf('Query name %s already exists in the Query Manager, please give this query a unique name', $query->getName()));
         }
 
-        // Check data provider
-        if (!$this->hasDataProvider($dataProviderName)) {
-            throw new QueryManagerException(sprintf('Data provider %s not found, please add this first via QueryManagerOld::addDataProvider()', $dataProviderName));
+        // Get data provider (pass as 2nd argument or use current data provider)
+        if ($dataProviderName === null) {
+            $dataProviderName = $this->getLastDataProviderName();
         }
         $dataProvider = $this->getDataProvider($dataProviderName);
-        if (!$query->checkDataProvider($dataProvider)) {
-            throw new QueryManagerException(sprintf('Data provider %s is not compatible with passed query %s', $dataProviderName, $queryName));
-        }
 
         // Prepare request
-        $query->prepareRequest($this->getDataProvider($dataProviderName));
+        switch (get_class($query)) {
+            case GraphQLQuery::class:
+                $buildQuery = new BuildGraphQLQuery($dataProvider);
+                break;
+            case Query::class:
+            default:
+                $buildQuery = new BuildQuery($dataProvider);
+                break;
+        }
 
-        $this->queries[$queryName] = $query;
-        $this->dataProvidersForQueries[$queryName] = $dataProviderName;
+        $response = $buildQuery->prepareRequest($query);
+
+        // Add query, data provider name & response to query stack
+        $this->queryStack->add($query->getName(), new StackItem($query, $dataProviderName, $response));
     }
 
     /**
@@ -120,32 +152,31 @@ class QueryManager
     {
         // Build array of concurrent queries for all responses that have not been executed
         $concurrent = [];
-        foreach ($this->getQueries() as $queryName => $query) {
-            if (!$query->hasResponse() || $query->hasResponseRun()) {
+        foreach ($this->getQueryStack() as $name => $item) {
+            if ($item->hasResponseRun()) {
                 continue;
             }
-            $concurrent[$queryName] = $query;
-        }
 
-        // Run multiple queries concurrently for performance
-        foreach ($concurrent as $queryName => $query) {
-            $dataProvider = $this->getDataProviderForQuery($queryName);
+            $query = $item->getQuery();
+            $dataProvider = $this->getDataProvider($item->getDataProviderName());
+            $response = $item->getResponse();
+
             if ($query->isSubRequest()) {
                 $dataProvider->suppressErrors();
             } else {
                 $dataProvider->suppressErrors(false);
             }
-            $dataProvider->runRequest($query->getResponse());
+            $dataProvider->runRequest($response);
         }
     }
 
     /**
      * Return all queries
-     * @return Query[]
+     * @return QueryStack Collection of queries
      */
-    public function getQueries(): array
+    public function getQueryStack(): QueryStack
     {
-        return $this->queries;
+        return $this->queryStack;
     }
 
     /**
@@ -155,7 +186,7 @@ class QueryManager
      */
     public function hasQuery(string $name): bool
     {
-        return (isset($this->queries[$name]));
+        return (isset($this->queryStack[$name]));
     }
 
     /**
@@ -163,10 +194,10 @@ class QueryManager
      * @param string $name
      * @return Query|null
      */
-    public function getQuery(string $name): ?Query
+    public function getQuery(string $name): ?StackItem
     {
         if ($this->hasQuery($name)) {
-            return $this->queries[$name];
+            return $this->queryStack[$name];
         }
         return null;
     }
@@ -179,16 +210,18 @@ class QueryManager
      */
     public function getResponse(string $queryName): CacheableResponse
     {
-        if (!$this->hasQuery($queryName)) {
+        if (!$this->queryStack->exists($queryName)) {
             throw new QueryManagerException(sprintf('Cannot find query with query name "%s"', $queryName));
         }
-        $query = $this->getQuery($queryName);
+
         $this->runQueries();
-        if (!$query->hasResponseRun()) {
+
+        $item = $this->queryStack->get($queryName);
+        if (!$item->hasResponseRun()) {
             throw new QueryManagerException(sprintf('Response has not run for query name "%s"', $queryName));
         }
 
-        return $query->getResponse();
+        return $item->getResponse();
     }
 
     /**
@@ -197,25 +230,31 @@ class QueryManager
      * Default functionality is to return decoded data as an array
      *
      * @param string $queryName
+     * @param string|null $rootPropertyPath Property path to root element to select data from
      * @throws QueryManagerException
      *@todo Create an Item object to return (which is cache aware)
      */
-    public function getItem(string $queryName): array
+    public function getItem(string $queryName, ?string $rootPropertyPath = null): array
     {
-        if (!$this->hasQuery($queryName)) {
+        if (!$this->queryStack->exists($queryName)) {
             throw new QueryManagerException(sprintf('Cannot find query with query name "%s"', $queryName));
         }
 
         // Run queries
-        $dataProvider = $this->getDataProviderForQuery($queryName);
         $this->runQueries();
-        $query = $this->getQuery($queryName);
-        if (!$query->hasResponseRun()) {
+
+        $item = $this->queryStack->get($queryName);
+        if (!$item->hasResponseRun()) {
             throw new QueryManagerException(sprintf('Response has not run for query name "%s"', $queryName));
         }
 
         // Return decoded data
-        return $dataProvider->decode($query->getResponse());
+        $dataProvider = $this->getDataProvider($item->getDataProviderName());
+        $query = $item->getQuery();
+        $data = $dataProvider->decode($item->getResponse());
+
+        $mapper = new MapItem(new WildcardMappingStrategy());
+        return $mapper->map($data, $rootPropertyPath);
     }
 
     /**
@@ -224,24 +263,27 @@ class QueryManager
      * Default functionality is to return decoded data as an array with pagination
      *
      * @param string $queryName
+     * @param string|null $rootPropertyPath Property path to root element to select data from
      * @throws QueryManagerException
      */
-    public function getCollection(string $queryName): Collection
+    public function getCollection(string $queryName, ?string $rootPropertyPath = null): Collection
     {
-        if (!$this->hasQuery($queryName)) {
+        if (!$this->queryStack->exists($queryName)) {
             throw new QueryManagerException(sprintf('Cannot find query with query name "%s"', $queryName));
         }
-        $query = $this->getQuery($queryName);
-        $dataProvider = $this->getDataProviderForQuery($queryName);
 
         // Run queries
         $this->runQueries();
-        if (!$query->hasResponseRun()) {
+
+        $item = $this->queryStack->get($queryName);
+        if (!$item->hasResponseRun()) {
             throw new QueryManagerException(sprintf('Response has not run for query name "%s"', $queryName));
         }
 
         // Return collections object with decoded data & pagination
-        $response = $query->getResponse();
+        $dataProvider = $this->getDataProvider($item->getDataProviderName());
+        $query = $item->getQuery();
+        $response = $item->getResponse();
         $data = $dataProvider->decode($response);
 
         $mapper = new MapCollection(new WildcardMappingStrategy());
@@ -252,7 +294,7 @@ class QueryManager
             $mapper->fromPaginationData($response->getHeaders());
         }
 
-        return $mapper->map($data);
+        return $mapper->map($data, $rootPropertyPath);
     }
 
 }