+++ CHANGELOG.md
@@ -9,3 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Refactored [Frontend](https://github.com/strata/frontend) data layer into its own repo
 - Flexible data access layer for accessing data, transforming, and mapping from HTTP APIs
+- Added support for GraphQL, Rest and HTTP data providers
+- Added query manager and query objects to help sending data requests and structure returned data
+-